requirements.txt
asgiref==3.4.1
autopep8==1.5.7
botocore==1.20.104
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
jmespath==0.10.0
Pillow==8.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
